Fix timeout error when approving request
Sending email, discord messages and reddit submissions worked, and then
the transaction timeouted.

Now, all possible entities are deleted in a single place, and the
timeout has been increased. Not only it will stop failing, when it does
it will correctly delete the discord messages and reddit posts. Sadly we
can't unsend an email