Fix readableTimeDiff Issues
readableTimeDiff would miscount the number of days difference depending on what the time of day the page was viewed and the time of day an application/request was sent.

For example if a support request was viewed at past midday on 2022/12/3 and the due date it was being compared to was 2022/12/5, the readableTimeDiff would incorrectly round to a 1 day difference.

Inversely, if a support request was viewed at past midday on 2022/12/3 and the due date it was being compared to was 2022/12/1, the readableTimeDiff would incorrectly round to a 3 day difference.

I have fixed this by setting both the current and comparison DateTime to the start of their respective days (0 hrs, 0 secs, 0 milliseconds) to stop their times being a factor.

Another issue was that readableTimeDiff returned [null,null] for any time difference lower than 24 hours which would be visualised as posted/due/applied 'Today'. This meant if you viewed a support request at 7am on 2022/12/3 that had a due date of 2022/12/4, the due date would be shown as 'Today'.

I have fixed this by instead having readableTimeDiff return [null,null] if the current and comparison DateTime have the same date.